+++ lib/strava/models/heart_rate_zone_ranges.rb
@@ -1,0 +1,8 @@
+module Strava
+  module Models
+    class HeartRateZoneRanges < Model
+      property 'custom_zones'
+      property 'zones', transform_with: ->(v) { v.map { |r| Strava::Models::ZoneRange.new(r) } }
+    end
+  end
+end