+++ lib/strava/models/power_zone_ranges.rb
@@ -1,0 +1,7 @@
+module Strava
+  module Models
+    class PowerZoneRanges < Model
+      property 'zones', transform_with: ->(v) { v.map { |r| Strava::Models::ZoneRange.new(r) } }
+    end
+  end
+end