+++ lib/strava/models/zone_range.rb
@@ -1,0 +1,8 @@
+module Strava
+  module Models
+    class ZoneRange < Model
+      property 'max'
+      property 'min'
+    end
+  end
+end